+++ src/types.ts
@@ -15,7 +15,7 @@ export enum Command {
 	SHOW_FLOATING_WINDOW = 'show_floating_window',
 	HIDE_FLOATING_WINDOW = 'hide_floating_window',
 	SET_NOTIFICATION = 'set_notification',
-	SET_ACTIVE_MODE = 'set_active_mode',
+	SET_FOCUS_MODE = 'set_focus_mode',
 }
 
 export enum Event {
@@ -158,7 +158,7 @@ export type Args<T extends Command> = {
 	[Command.SET_NOTIFICATION]: {
 		number?: number;
 	};
-	[Command.SET_ACTIVE_MODE]: boolean;
+	[Command.SET_FOCUS_MODE]: boolean;
 }[T];
 
 export type CommandResponse<T extends Command> = {
@@ -180,7 +180,7 @@ export type CommandResponse<T extends Command> = {
 	[Command.SHOW_FLOATING_WINDOW]: void;
 	[Command.HIDE_FLOATING_WINDOW]: void;
 	[Command.SET_NOTIFICATION]: void;
-	[Command.SET_ACTIVE_MODE]: void;
+	[Command.SET_FOCUS_MODE]: void;
 }[T];
 
 export type MessageChannelCommandResponse<T extends Command> = {