+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - New `SET_NOTIFICATION` command
-- New `SET_ACTIVE_MODE` command
+- New `SET_FOCUS_MODE` command
 
 ## [0.4.0] - 2022-10-31